refactor: move alignment results to avoid circular import
